Reorganized vaidation tests.
